chore(paw): story 7-6-1 progressed to dev-story [Story-7-6-1]
- Quality gate passed (`./ctl check` reports `✓ Quality gate passed`). The errors shown in the output are from cppcheck running on macOS without the PCH context — those are expected and don't affect the quality gate result.
Story 7-6-1 dev-story phase complete. Here's the summary:

+++ src/source/Gameplay/Buffs/w_BuffTimeControl.cpp
@@ -6,6 +6,8 @@
 #include "UIManager.h"
 #include "ItemAddOptioninfo.h"
 #include "w_BuffTimeControl.h"
+#include "w_BuffScriptLoader.h"
+#include "_GlobalFunctions.h"
 
 BuffTimeControlPtr BuffTimeControl::Make()
 {
@@ -93,7 +95,7 @@ void BuffTimeControl::RegisterBuffTime(eBuffState bufftype, DWORD curbufftime)
 
     curbufftime = GetBuffMaxTime(bufftype, curbufftime);
 
-    if (bufftimetype == eBuffTime_None || curbufftime == -1)
+    if (bufftimetype == eBuffTime_None || curbufftime == static_cast<DWORD>(-1))
     {
         return;
     }
